requirements.txt
black      ==         23.11.0
djangorestframework == 3.14.0
psycopg2-binary  ==   2.9.9




